+++ OpenHABCore/Sources/OpenHABCore/Model/OpenHABWidget.swift
@@ -1,4 +1,4 @@
-// Copyright (c) 2010-2025 Contributors to the openHAB project
+// Copyright (c) 2010-2026 Contributors to the openHAB project
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information.
@@ -75,9 +75,9 @@ public class OpenHABWidget: NSObject, MKAnnotation, Identifiable, ObservableObje
     public var refresh = 0
     public var height: Double?
     public var isLeaf = false
-    public var iconColor = ""
-    public var labelcolor = ""
-    public var valuecolor = ""
+    @Published public var iconColor = ""
+    @Published public var labelcolor = ""
+    @Published public var valuecolor = ""
     public var service = ""
     @Published public var state = ""
     public var text = ""