+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## [Unreleased]
 
+## [Version 3.1.45, Build 102] - 2026-01-19Z
+
+- Fix for race condition in NetworkTrackerTests (#1044)
+- Fix/1038 auto back from selection view (#1041)
+- Replaced print statements with Logger calls
+
+## [Version 3.1.44, Build 101] - 2026-01-15Z
+
+- Implement SVG color preprocessing for non-iconify icons in OpenHABImageProcessor (#1036)
+- Add PR review instructions
+- Update to 2026 Copyright message (#1035)
+
+## [Version 3.1.43, Build 100] - 2026-01-01Z
+
+- Clean up obsolete code after migration to openapi-generator (#1032)
+
 ## [Version 3.1.42, Build 99] - 2025-12-24Z
 
 - Let the webview scroll its overflow under the status bar, similar to previous versions (#1031)